Move position classes to separate folder
