--- game/classes.py
@@ -1,255 +1,0 @@
-import math
-
-
-class Pos2(object):
-    def __init__(self, x=0, y=0):
-        super(Pos2, self).__init__()
-
-        self.x = x
-        self.y = y
-
-
-    def clamp(self, minpos, maxpos):
-        self.x = max(min(self.x, maxpos.x), minpos.x)
-        self.y = max(min(self.y, maxpos.y), minpos.y)
-    
-    def loop_around(self, maxpos):
-        if self.x >= maxpos.x: self.x %= maxpos.x
-        if self.y >= maxpos.y: self.y %= maxpos.y
-    
-    def distancesq_to(self, pos):
-        return (self.x - pos.x)**2 + (self.y - pos.y)**2
-    
-    def angle_to(self, pos):
-        x = self.x - pos.x
-        y = self.y - pos.y
-        return -math.degrees(math.atan2(y, x))
-
-    @classmethod
-    def from_pos2(cls, pos2):
-        return cls(*pos2.to_coords())
-    
-    @classmethod
-    def from_pos3(cls, pos3):
-        return cls(*pos3.to_coords()[0:2])
-
-    def to_coords(self):
-        return self.x, self.y
-    
-    @classmethod
-    def from_str(cls, s):
-        return cls(*eval(s))
-
-    def to_list(self):
-        return [self.x, self.y]
-
-    @classmethod
-    def from_list(cls, lst):
-        return cls(*lst)
-
-    def to_dict(self):
-        return {"x": self.x, "y": self.y}
-
-    @classmethod
-    def from_dict(cls, dct):
-        return cls(**dct)
-
-
-    def __bool__(self):
-        return self.x != 0 or self.y != 0
-
-    def __eq__(self, other):
-        try:
-            return self.x == other.x and self.y == other.y
-        except:
-            return self.x == other and self.y == other
-    
-    def __add__(self, other):
-        try:
-            return self.__class__(self.x + other.x, self.y + other.y)
-        except:
-            return self.__class__(self.x + other, self.y + other)
-
-    def __sub__(self, other):
-        try:
-            return self.__class__(self.x - other.x, self.y - other.y)
-        except:
-            return self.__class__(self.x - other, self.y - other)
-    
-    def __mul__(self, other):
-        try:
-            return self.__class__(self.x * other.x, self.y * other.y)
-        except:
-            return self.__class__(self.x * other, self.y * other)
-    
-    def __truediv__(self, other):
-        try:
-            return self.__class__(self.x / other.x, self.y / other.y)
-        except:
-            return self.__class__(self.x / other, self.y / other)
-    
-    def __floordiv__(self, other):
-        try:
-            return self.__class__(self.x // other.x, self.y // other.y)
-        except:
-            return self.__class__(self.x // other, self.y // other)
-    
-    def __mod__(self, other):
-        try:
-            return self.__class__(self.x % other.x, self.y % other.y)
-        except:
-            return self.__class__(self.x % other, self.y % other)
-    
-    def __round__(self):
-        return self.__class__(round(self.x), round(self.y))
-    
-    def __trunc__(self):
-        return self.__class__(math.trunc(self.x), math.trunc(self.y))
-
-    def __str__(self):
-        return f"({self.x}, {self.y})"
-
-
-class Pos3(object):
-    def __init__(self, x=0, y=0, z=0):
-        super(Pos3, self).__init__()
-
-        self.x = x
-        self.y = y
-        self.z = z
-
-
-    def clamp(self, minpos, maxpos):
-        self.x = max(min(self.x, maxpos.x), minpos.x)
-        self.y = max(min(self.y, maxpos.y), minpos.y)
-        self.z = max(min(self.z, maxpos.z), minpos.z)
-
-
-    @classmethod
-    def from_pos3(cls, pos3):
-        return cls(*pos3.to_coords())
-
-    def to_coords(self):
-        return self.x, self.y, self.z
-    
-    @classmethod
-    def from_str(cls, s):
-        return cls(*eval(s))
-    
-    def to_list(self):
-        return [self.x, self.y, self.z]
-
-    @classmethod
-    def from_list(cls, lst):
-        return cls(*lst)
-
-    def to_dict(self):
-        return {"x": self.x, "y": self.y, "z": self.z}
-
-    @classmethod
-    def from_dict(cls, dct):
-        return cls(**dct)
-
-
-    def __bool__(self):
-        return self.x != 0 or self.y != 0 or self.z != 0
-
-    def __eq__(self, other):
-        try:
-            return self.x == other.x and self.y == other.y and self.z == other.z
-        except: pass
-        try: 
-            return self.x == other.x and self.y == other.y and self.z == 0
-        except: pass
-        return self.x == other and self.y == other and self.z == other
-    
-    def __add__(self, other):
-        try:
-            return self.__class__(self.x + other.x, self.y + other.y, self.z + other.z)
-        except: pass
-        try: 
-            return self.__class__(self.x + other.x, self.y + other.y, self.z)
-        except: pass
-        return self.__class__(self.x + other, self.y + other, self.z + other)
-    
-    def __sub__(self, otherpos):
-        try:
-            return self.__class__(self.x - otherpos.x, self.y - otherpos.y, self.z - otherpos.z)
-        except: pass
-        try: 
-            return self.__class__(self.x - other.x, self.y - other.y, self.z)
-        except: pass
-        return self.__class__(self.x - other, self.y - other, self.z - other)
-    
-    def __mul__(self, other):
-        try:
-            return self.__class__(self.x * other.x, self.y * other.y, self.z * other.z)
-        except: pass
-        try: 
-            return self.__class__(self.x * other.x, self.y * other.y, self.z)
-        except: pass
-        return self.__class__(self.x * other, self.y * other, self.z * other)
-    
-    def __truediv__(self, other):
-        try:
-            return self.__class__(self.x / other.x, self.y / other.y, self.z / other.z)
-        except: pass
-        try: 
-            return self.__class__(self.x / other.x, self.y / other.y, self.z)
-        except: pass
-        return self.__class__(self.x / other, self.y / other, self.z / other)
-    
-    def __floordiv__(self, other):
-        try:
-            return self.__class__(self.x // other.x, self.y // other.y, self.z // other.z)
-        except: pass
-        try: 
-            return self.__class__(self.x // other.x, self.y // other.y, self.z)
-        except: pass
-        return self.__class__(self.x // other, self.y // other, self.z // other)
-        
-    def __mod__(self, other):
-        try:
-            return self.__class__(self.x % other.x, self.y % other.y, self.z % other.z)
-        except: pass
-        try: 
-            return self.__class__(self.x % other.x, self.y % other.y, self.z)
-        except: pass
-        return self.__class__(self.x % other, self.y % other, self.z % other)
-    
-    def __int__(self):
-        return self.__class__(int(self.x), int(self.y), int(self.z))
-
-    def __round__(self):
-        return self.__class__(round(self.x), round(self.y), round(self.z))
-    
-    def __trunc__(self):
-        return self.__class__(math.trunc(self.x), math.trunc(self.y), math.trunc(self.z))
-
-    def __str__(self):
-        return f"({self.x}, {self.y}, {self.z})"
-
-
-class Vector2(object):
-    def __init__(self, x, y, *args, **kwargs):
-        super(Vector2, self).__init__(*args, **kwargs)
-        
-        self.x = x
-        self.y = y
-    
-    @classmethod
-    def ZERO(cls):
-        return cls(0.0, 0.0)
-    
-    def normalize(self, precision=3):
-        v = math.atan2(self.y, self.x)
-        self.x = round(math.cos(v), precision)
-        self.y = round(math.sin(v), precision)
-
-    
-    def __bool__(self):
-        return self.x != 0 or self.y != 0
-
-    def __str__(self):
-        return f"({self.x}, {self.y})"
-
