+++ game/positions/__init__.py
@@ -1,0 +1,2 @@
+from .classes import *
+from .positions import *
